requirements.txt
# UI
streamlit>=1.30
pywebview>=4.4
streamlit-drawable-canvas
streamlit-autorefresh
streamlit-image-coordinates>=0.4.0
# Numerics / ML
numpy>=1.26
pandas>=2.0
scikit-learn>=1.4
joblib>=1.3
pyarrow>=14.0
st-cytoscape>=0.0.5
# Computer vision & images
opencv-python>=4.8
pillow>=10.0
st-cytoscape

# Nice-to-have utilities (safe)
tqdm>=4.66
